Read-Only Usage Interfaces
Not sure if this totally makes sense, but I feel like it does. These should be usable readonly, even if they are mutable at runtime.

I was going to make the Options interface readonly too, but I think that might be an anti-pattern, not totally sure. I haven't seen a readonly Options config before to confirm that it should be readonly.